complate Login (iOS8)

+++ FacebookSample/FacebookSample.iOS/ExWebViewRenderer.cs
@@ -8,25 +8,20 @@ using Xamarin.Forms.Platform.iOS;
 namespace FacebookSample.iOS {
     public class ExWebViewRenderer : WebViewRenderer {
 
-
         protected override void OnElementChanged(VisualElementChangedEventArgs e) {
             base.OnElementChanged(e);
 
             //Xamarin.Formのコントロール(ExWebView)
             var exWebView = e.NewElement as ExWebView;
             //ネイティブコントロール(MonoTouch.UIKit.UIWebView)
-            var webView = this;
+            //var webView = this;
 
-            webView.ShouldStartLoad = (w, request, naviType) => {
-                //イベントをForms側に送る
-                exWebView.OnNavigate(request.Url.AbsoluteString);
-                return true;
-            };
-
-            //クッキー（ログイン情報）の削除
-            var storage = NSHttpCookieStorage.SharedStorage;
-            foreach (var cookie in storage.Cookies) {
-                storage.DeleteCookie(cookie);
+            if (exWebView.DeleteCookie) {
+                //クッキー（ログイン情報）の削除
+                var storage = NSHttpCookieStorage.SharedStorage;
+                foreach (var cookie in storage.Cookies) {
+                    storage.DeleteCookie(cookie);
+                }
             }
 
         }
